Validatio of stories count added. Stories count must be greater than zero
